[REFACTOR]: moved service/reducers/middleware into addon directory
