+++ addon/middleware/index.js
@@ -1,0 +1,5 @@
+import thunk from 'redux-thunk';
+
+var resolved = thunk.default ? thunk.default : thunk;
+
+export default [ resolved ];
