+++ app/enhancers/index.js
@@ -1,1 +1,1 @@
-export {default} from 'ember-redux/enhancers'
+export {default} from 'ember-redux/enhancers/index';
