+++ app/middleware/index.js
@@ -1,5 +1,1 @@
-import thunk from 'redux-thunk';
-
-var resolved = thunk.default ? thunk.default : thunk;
-
-export default [ resolved ];
+export {default} from 'ember-redux/middleware/index';
